# Avoid some uses of Triangulation::user_flags