While working on the internal restructuring of `DoFHandler`, and also when looking at performance profiles, I observed an annoyingly high number of calls to `Triangulation::save_user_flags` and `Triangulation::load_user_flags`. We used those in many places to store some information regarding the lines or faces of a triangulation we only want to visit once, e.g. during the enumeration of the DoFs. However, that comes with a lot of boilerplate, because we always need to save the user flags, do our thing, and restore them. Also from a performance point of view, these operations are not for free, because `std::vector<bool>` has no single-instruction access function.

This PR cleans up the uses outside the immediate grid constructions that should be safe: Keeping a separate vector with the flags communicates the intent much more clearly, and avoids going back and forth between a user's flags, thus reducing the time in those setup functions. A net of 270 reduced lines of code shows that this is a significant simplification. (OK, 20 lines or so are due to range-based for loops I changed on the way).

I tagged this with the 9.4 milestone because I want to get it done, but it is not an urgent PR.