@bangerth 2022-05-30
This is the sort of thing that makes me nervous just before a release, at least the parts that aren't in `dof_handler.cc` and `tria.cc`. I think the patch is good, but since it does not provide functionality we didn't have before my vote would be to let the patch wait until after we create the branch. Do you insist on having it in 9.4? If you do, could you break it into those parts that are definitely tested in hundreds of tests (e.g., in `DoFHandler::reserve_*()`) which I'd be ok to merge, and those parts that maybe don't get run in half of our tests?