@kronbichler 2022-05-30
For the record, while creating the patch, the combination of hp and multigrid tests that I ran to verify did uncover three places where I had made a mistake, so we seem to indeed have covered most of these places reasonably well.