@kronbichler 2022-06-03
For reference, this PR will have significant effect on our performance tests. The `timing_step37`, when I run it on 1 node / 72 MPI ranks, I record 0.77 seconds for the subtest `setup_dofs` on top of this PR and #13828, compared to 1.00 seconds on current master. We do regress compared to the beginning of May since the benchmark added DoF renumbering function via #13674 (when the time was around 0.4 seconds), but we are significantly ahead of the 0.94 seconds we had before I started my cleanups in DoFHandler and related two weeks ago.

Regarding larger scales, we are significantly ahead already before this PR: On 64 nodes / 4608 MPI ranks, the `create_grid` function went from 0.22 seconds to 0.11 seconds on an average of 20 samples, and the `setup_dofs` function went from 0.062 to 0.055 seconds (so despite being slower on 1 node than we were in mid May; we slightly improve on this number, too, despite this being inside the scaling limit).